requirements.txt
fastapi
uvicorn
email-validator
gspread
google-auth
openai